+++ CHANGELOG.md
@@ -1,6 +1,12 @@
 Changelog
 =========
 
+* **2015-02-11** It is no longer possible to change the event dispatcher of the
+  CacheInvalidator once its instantiated. If you need a custom dispatcher, set it
+  right after creating the invalidator instance. Deprecated CacheInvalidator::addSubscriber
+  in favor of either using the event dispatcher instance you inject or do
+  getEventDispatcher()->addSubscriber($subscriber)
+
 1.2.0
 -----